Updated recommendation.js with time functionality
After the initial loading, the pop-up will display the current recommendation until the chosen time expires. Then the computeRecommendation function is called to compute a new recommendation. This is done when the pop-up is open.
When the pop-up isn't open, recommendation.js sends a message to background.js. When the response is received, computeRecommendation is called, resulting in a new recommendation when the pop-up is opened.