Filter functionality implemented
Implemented filter functionality.
Operates on the logic where if an anime matches at least one of the filters, it will be included in the recommendation. Am working on implementing functionality where recommendation must meet all filters chosen.
Also, the settings page doesn't remember which filters were checked in a previous section, so recalling of previously checked boxes needs to be implemented.