Adjusted delayInMinutes and added clear
delayInMinutes is the initial first "alarm" that goes off in X amount of minutes. The times have been adjusted to the correct times for each. After the first "alarm" goes off, another is set to go off at periods of the same time later. Each button now has a "clear", which removes any instance of the alarm named "time" (which they are all named).