Have the page open with the saved font size
Previously, the user value for the font size was already saved to localStorage. However, when the page was closed and reopened, the page would revert back to the default font size. Changed it so that the page is always displayed in the correct font size by retrieving the saved value from localStorage and updating the page on load.